# Conflicts:
#	src/main/java/br/ufg/inf/pitanga/controller/SistemaController.java
#	src/main/java/br/ufg/inf/pitanga/entidades/Ingresso.java

+++ src/main/java/br/ufg/inf/pitanga/dao/IngressoDao.java
@@ -41,7 +41,7 @@ public class IngressoDao implements InterfaceIngressoDao {
     }
 
     @Override
-    public Ingresso buscarPorId(int id) {
+    public Ingresso buscarPorId(Long id) {
         return manager.find(Ingresso.class, id);
     }
